refactor(wal): make the ring the sole WAL backend; remove legacy channel+worker
Per the measured result (ring beats RocksDB 2.8-5.5x on concurrent durable
write at 1/4/8/16 threads after the hardening) and full validation, the
lock-free shared WAL ring replaces the legacy channel+worker outright.

- Delete src/journal/group_commit.rs (legacy channel+worker Journal).
- Rename group_commit_ring.rs -> group_commit.rs (the ring is now THE
  coordinator); move JournalStats into it.
- Drop the `wal_ring` cargo feature (root + benches) and all cfg gating;
  `ring` + `group_commit` are compiled unconditionally. journal/mod.rs
  re-exports the single Journal.
- Doc cleanup: ring.rs / wal-ring.md / PERF_FINDINGS reflect "sole backend,
  legacy removed" (no flag).

Validation, ring as DEFAULT, both arches: full lib 279 + properties (proptest
BTreeMap/WAL-replay oracle) 5 + checkpoint_failpoint 8 + concurrent_stress 3 +
wal_tree_integration 39 + tree_smoke 103 + snapshot 13 + batch/checkpoint/
scan_stats + loom gap-safety (2+3 publishers) 2; clippy clean all-targets;
multi-process SIGKILL crash-soak 40 rounds (every recovery a contiguous valid
prefix).
