+++ src/journal/mod.rs
@@ -22,26 +22,15 @@
 //! tree + journal boundary.
 
 pub mod codec;
+// The WAL append coordinator: a lock-free shared ring (`ring`) + a single
+// flusher (`group_commit`). See docs/design/wal-ring.md.
 pub(crate) mod group_commit;
 pub mod reader;
+pub(crate) mod ring;
 pub mod wal_op;
 pub mod writer;
 
-// Lock-free shared WAL ring (docs/design/wal-ring.md): the ring buffer
-// (`ring`) + the ring-backed coordinator (`group_commit_ring`). Behind the
-// `wal_ring` feature; the default build uses the legacy channel+worker.
-#[cfg(feature = "wal_ring")]
-pub(crate) mod ring;
-#[cfg(feature = "wal_ring")]
-pub(crate) mod group_commit_ring;
-
-/// The active WAL coordinator: ring-backed under `wal_ring`, else the legacy
-/// channel+worker. Both expose the same `pub(crate)` API and share the
-/// internal `JournalStats` type from `group_commit`.
-#[cfg(not(feature = "wal_ring"))]
 pub(crate) use group_commit::Journal;
-#[cfg(feature = "wal_ring")]
-pub(crate) use group_commit_ring::Journal;
 
 #[cfg(test)]
 mod tests;
